fix: Node.delete() should not delete content_objects from other apps

+++ business_logic/models/node.py
@@ -42,10 +42,16 @@ class Node(NS_Node):
         return super(Node, cls).add_root(**kwargs)
 
     def delete(self):
-        if self.object_id and self.content_object:
+        if (
+            self.object_id and
+            self.content_object and
+            self.content_type.app_label == ContentType.objects.get_for_model(self.__class__).app_label
+        ):
             self.content_object.delete()
+
         for child in self.get_children():
             child.delete()
+
         return super(Node, self).delete()
 
     def add_child(self, **kwargs):
